% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{covid_uk_data}
\alias{covid_uk_data}
\title{Covid-19 UK data}
\description{
Contains the weights of individual models for each quantile as estimated in
the different variants (qra_model) of quantile regression averaging (QRA).
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
