% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{ensembles}
\alias{ensembles}
\title{UK ensemble forecasts in quantile format}
\description{
Contains ensembles of Covid-19 forecasts made between 31 March and 13 July 2020, by geography (NHS region and UK nation), value_type (code for different targets), value_desc (description of the traget), creation_date (the date on which the forecast was made), value_date (the predicted date), quantile (the predictive quantile) and value (the forecast at the given quantile); ensembles have been created as equally weighted quantiles (EWQ) and using a quantile regression average (QRA), using the `generate_ensembles` function and encompassing a variety of model variants.
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
