% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_quantiles.r
\name{estimate_quantiles}
\alias{estimate_quantiles}
\title{estimate missing quantiles}
\usage{
estimate_quantiles(x)
}
\arguments{
\item{x}{a vector of quantiles, with levels as names and some missing (NA) values}
}
\value{
a vector of quantiles, with levels as names and all values
}
\description{
estimate missing quantiles
}
\author{
Sebastian Funk
}
