% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewq_rank_table.r
\name{ewq_rank_table}
\alias{ewq_rank_table}
\title{Formats a table of ranked QRA ensembles}
\usage{
ewq_rank_table(ranked_ensembles, file, format = "latex", ...)
}
\arguments{
\item{file}{filename, if to be saved into a file}

\item{format}{table format}

\item{...}{any further parameters to pass to \code{kable}}

\item{ensembles}{data frame with ranked ensembles}
}
\value{
table
}
\description{
Formats a table of ranked QRA ensembles
}
\author{
Sebastian Funk
}
