% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{exclude_data}
\alias{exclude_data}
\title{Data to exclude from scoring}
\description{
Contains data sets to exclude from scoring because the they were not available at the time the forecasts were made.
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
