% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_null_model.r
\name{fit_null_model}
\alias{fit_null_model}
\title{Create null model forecasts}
\usage{
fit_null_model(forecasts, data)
}
\arguments{
\item{forecasts}{existing forecast data frame}

\item{data}{data to create null model}
}
\value{
data frame with null model forecasts
}
\description{
Uses a truncated normal distribution fitted to past data points under the
assumption of no change. Contains uk-specific code on truncation at the time
of data availability
}
\author{
Sebastian Funk
}
