% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ensembles.r
\name{generate_ensembles}
\alias{generate_ensembles}
\title{}
\usage{
generate_ensembles(forecasts, data, all_dates = TRUE, max_history = 5, ...)
}
\arguments{
\item{forecasts}{data frame of forecasts}

\item{data}{data from of validation data}

\item{all_dates}{whether ensembles are to be generated for all creation
dates; otherwise will just do the latest}

\item{max_history}{maximum history to consider in QRA}

\item{...}{any options to pass to the \link{qra} function.}
}
\value{
a data frame with the ensembles
}
\description{
Generate quantile regression ensembles and an equal-weighted ensemble
}
\details{
Produces variants of QRA where the weights are optimised per quantile or
not, with or without intercept, and normalised to sum up to 1 or not.
}
\author{
Sebastian Funk
}
