% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{nhse_regions}
\alias{nhse_regions}
\title{NHSE regions}
\description{
A character vector containing the names of National Health Service England
(NHSE) regions.
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
