% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_null_model.r
\name{null_model_forecast_quantiles}
\alias{null_model_forecast_quantiles}
\title{Makes a null model forecast based on a vector of values}
\usage{
null_model_forecast_quantiles(values, horizon, truncation = 0)
}
\arguments{
\item{values}{vector of values}

\item{horizon}{forecast horizon}

\item{truncation}{truncation of data series}
}
\value{
quantiles of null model predictive distribution
}
\description{
Null model assumes last value stays
}
\author{
Sebastian Funk
}
