% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_plot.r
\name{null_model_plot}
\alias{null_model_plot}
\title{Plot performance against the null model}
\usage{
null_model_plot(forecasts, null, horizons = c(7, 14))
}
\arguments{
\item{forecasts}{a list of data frames with scored forecasts}

\item{null}{data frame with scored null model}

\item{horizons}{forecast horizons to consider}

\item{data}{data frame with data}
}
\value{
ggplot object
}
\description{
Plot performance against the null model
}
\author{
Sebastian Funk
}
