% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_plot.r
\docType{data}
\name{p}
\alias{p}
\title{Plot performance against the null model}
\format{
An object of class \code{gg} (inherits from \code{ggplot}) of length 9.
}
\usage{
p
}
\arguments{
\item{forecasts}{a list of data frames with scored forecasts}

\item{data}{data frame with data}

\item{null}{data frame with scored null model}

\item{horizons}{forecast horizons to consider}
}
\value{
ggplot object
}
\description{
Plot performance against the null model
}
\author{
Sebastian Funk
}
\keyword{datasets}
