% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_forecasts.r
\name{plot_data_forecasts}
\alias{plot_data_forecasts}
\title{}
\usage{
plot_data_forecasts(forecasts, data, horizon = 7, uncertainty = TRUE)
}
\arguments{
\item{forecasts}{}

\item{horizons}{}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Sebastian Funk
}
