% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_ensembles.r
\name{rank_forecasts}
\alias{rank_forecasts}
\title{Rank forecasts by calculating an average across targets}
\usage{
rank_forecasts(x)
}
\arguments{
\item{x}{a data frame with scored forecasts}
}
\value{
a data frame with ranked forecasts, and mean/sd of the weighted interval score
}
\description{
Rank forecasts by calculating an average across targets
}
\author{
Sebastian Funk
}
