% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_forecasts.r
\name{score_forecasts}
\alias{score_forecasts}
\title{Score forecasts according to the weighted interval score and coverage}
\usage{
score_forecasts(forecast, data)
}
\arguments{
\item{forecast}{a data frame of forecasts}

\item{data}{a data frame of data with the same \code{value_date},
\code{geography}, \code{value_type} and \code{value_desc} entries as
\code{forecast}}
}
\value{
a data frame of forecasts with additional columns \code{score} (WIS)
and \code{inside} (whether the data are inside or outside the given interval
}
\description{
Calculates the weighted interval score for each value/data pair, and sets a
flag whether the data is inside (1) or outside the given interval
}
\author{
Sebastian Funk
}
