% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_table.r
\name{scores_table}
\alias{scores_table}
\title{Creates a table of forecast scores}
\usage{
scores_table(
  forecasts,
  file,
  coverage_levels = c(0.5, 0.9),
  horizons = c(7, 14),
  common_creation_dates = FALSE,
  format = "latex",
  ...
)
}
\arguments{
\item{forecasts}{data frame with scored forecasts}

\item{file}{filename, if to be saved into a file}

\item{coverage_levels}{coverage levels for assessing calibration}

\item{horizons}{forecast horizons to assess}

\item{common_creation_dates}{if TRUE, ensure only creation dates are assessed}

\item{format}{table format}

\item{...}{any further parameters to pass to \code{kable}}
}
\value{
table
}
\description{
Creates a table of forecast scores
}
\author{
Sebastian Funk
}
