% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_plot.r
\name{weights_plot}
\alias{weights_plot}
\title{Plot model weights from an ensemble}
\usage{
weights_plot(weights)
}
\arguments{
\item{weights}{a data frame containing the models and weights}

\item{plot_quantile}{the quantile to plot the weights for}
}
\value{
a ggplot object
}
\description{
Plot model weights from an ensemble
}
\author{
Sebastian Funk
}
