% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ensembles.r
\name{generate_ensembles}
\alias{generate_ensembles}
\title{}
\usage{
generate_ensembles(forecasts, data)
}
\arguments{
\item{forecasts}{data frame of forecasts}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Sebastian Funk
}
