% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_forecasts.r
\name{plot_data_forecasts}
\alias{plot_data_forecasts}
\title{Plot data and forecasts}
\usage{
plot_data_forecasts(
  forecasts,
  data,
  horizon = 7,
  uncertainty = TRUE,
  scales = "free_y",
  exclude = NULL
)
}
\arguments{
\item{forecasts}{data frame with forecasts}

\item{data}{truth data matching the forecasts}

\item{horizon}{forecast horizon to plot}

\item{uncertainty}{whether to plot uncertainty or not}

\item{scales}{any scales argument to pass to facet_wrap}

\item{exclude}{any data to exclude}

\item{horizons}{data frame with data}
}
\value{
ggplot object
}
\description{
Plot data and forecasts
}
\author{
Sebastian Funk
}
